move insertSnippet and toggleCodeGeneration to lib/testFile
